Alter NBTDataHandler code to deal with multiple dimensions
NBTDataHandler now holds an ArrayList of VillagerData objects to store
separate chunk data for each dimension.  init() is called after the server
has started to load or create data for each dimension.  The getters and
setters for the class have also been changed to require a dimensionID so we
can find the correct VillagerData.