Change TickHandler to tick dimensions one at a time
Lots of small changes here.  Class TickTracker added to keep track of
players on a per-world basis.  Logic added to find the next world to tick
and the next player to tick in that world.  Add logic to TickHandler to
track dimensions ticked using nextWorldToTick.  Add tracking for
TotalWorldTime to stop the code from ticking sequential dimensions all at
once if multiple players are present.