Rebuild SpawnHandler to work with chunkInhabitedTime
Because using getClampedAdditionalDifficulty gives weird quirks to the mod
like never spawning villagers on Peaceful difficulty and immediately starting
on Hard, we're going to stop using that.  It has been replaced by a simple
clamp of chunkInhabitedTime.

Also added:  Config options for the upper bound on the clamp, and to disable
or enable log messages for the mod.