Alter entitySpawned event handler to compare classes for mod compatibility
We used to check the name of an entity for a villager match - now we just
check if it's an instance of EntityVillager.  This should stop the code
from breaking on things that don't overwrite getName() for some reason and
potentially add mod compatibility for things like MCA if they use a
subclass of EntityVillager for their changes.